+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 17.06.2-ce (2017-09-12)
 - BUGFIX: auto detection of user proxy pac fixed
 - BUGFIX: BOOT2DOCKER_EXTENSION_DIR env var not available for proxyd fixed (because sudo)
+- BUGFIX: try to have something in http_proxy for user proxy pac in every context
 
 ## 17.06.2-ce (2017-09-06)
 - UPDATE: update to boot2docker 17.06.2-ce